order changing status

+++ server/HeatDev/Options/AuthOptions.cs
@@ -16,7 +16,7 @@ namespace HeatDev.Options
         public string Key { get; set; }
 
         public static readonly int REFRESH_LIFETIME = 180 * 24 * 60;
-        public static readonly double ACCESS_LIFETIME = 15;
+        public static readonly int ACCESS_LIFETIME = 15;
 
         public SymmetricSecurityKey GetSymmetricSecurityKey()
         {
